Improved $ref handling
All references can be handled the same way, no matter
if the spec is Swagger or Open API.

This commit also fixes the issue that references don't
work in referenced files. We no longer require that
a reference must exist in the main spec file.

Fixes #134
