+++ src/main/java/org/zalando/intellij/swagger/reference/OpenApiConstants.java
@@ -1,0 +1,9 @@
+package org.zalando.intellij.swagger.reference;
+
+public class OpenApiConstants {
+
+    public static final String SLASH = "/";
+    public static final String HASH = "#";
+    public static final String REFERENCE_PREFIX = "#/";
+    public static final String REF_KEY = "$ref";
+}
