+++ src/main/java/org/zalando/intellij/swagger/reference/SwaggerConstants.java
@@ -1,0 +1,12 @@
+package org.zalando.intellij.swagger.reference;
+
+public class SwaggerConstants {
+
+    public static final String SWAGGER = "swagger";
+    public static final String SLASH = "/";
+    public static final String REFERENCE_PREFIX = "#/";
+    public static final String HASH = "#";
+
+    public static final String TAGS_KEY = "tags";
+    public static final String REF_KEY = "$ref";
+}
