--- src/main/java/org/zalando/intellij/swagger/reference/swagger/OpenApiConstants.java
@@ -1,19 +1,0 @@
-package org.zalando.intellij.swagger.reference.swagger;
-
-public class OpenApiConstants {
-
-    public static final String SLASH = "/";
-    public static final String HASH = "#";
-    public static final String REFERENCE_PREFIX = "#/";
-    public static final String COMPONENT_REFERENCE_PREFIX = "#/components/";
-    public static final String REF_KEY = "$ref";
-
-    public static final String LOCAL_SCHEMAS_PREFIX = "#/components/schemas/";
-    public static final String LOCAL_RESPONSES_PREFIX = "#/components/responses/";
-    public static final String LOCAL_PARAMETERS_PREFIX = "#/components/parameters/";
-    public static final String LOCAL_EXAMPLES_PREFIX = "#/components/examples/";
-    public static final String LOCAL_REQUEST_BODIES_PREFIX = "#/components/requestBodies/";
-    public static final String LOCAL_HEADERS_PREFIX = "#/components/headers/";
-    public static final String LOCAL_LINKS_PREFIX = "#/components/links/";
-    public static final String LOCAL_CALLBACKS_PREFIX = "#/components/callbacks/";
-}
