--- src/main/java/org/zalando/intellij/swagger/reference/swagger/SwaggerConstants.java
@@ -1,18 +1,0 @@
-package org.zalando.intellij.swagger.reference.swagger;
-
-public class SwaggerConstants {
-
-    public static final String SWAGGER = "swagger";
-    public static final String SLASH = "/";
-    public static final String REFERENCE_PREFIX = "#/";
-    public static final String HASH = "#";
-
-    public static final String PARAMETERS_KEY = "parameters";
-    public static final String RESPONSES_KEY = "responses";
-    public static final String TAGS_KEY = "tags";
-    public static final String REF_KEY = "$ref";
-
-    public static final String LOCAL_DEFINITIONS_PREFIX = "#/definitions/";
-    public static final String LOCAL_PARAMETERS_PREFIX = "#/parameters/";
-    public static final String LOCAL_RESPONSES_PREFIX = "#/responses/";
-}
