+++ src/main/resources/META-INF/plugin.xml
@@ -23,11 +23,8 @@
         <completion.contributor language="JSON" implementationClass="org.zalando.intellij.swagger.completion.contributor.openapi.OpenApiJsonCompletionContributor"/>
         <completion.contributor language="yaml" implementationClass="org.zalando.intellij.swagger.completion.contributor.openapi.OpenApiYamlCompletionContributor"/>
 
-        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.contributor.swagger.SwaggerYamlReferenceContributor"/>
-        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.contributor.swagger.SwaggerJsonReferenceContributor"/>
-
-        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.contributor.openapi.OpenApiJsonReferenceContributor"/>
-        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.contributor.openapi.OpenApiYamlReferenceContributor"/>
+        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.YamlReferenceContributor"/>
+        <psi.referenceContributor implementation="org.zalando.intellij.swagger.reference.JsonReferenceContributor"/>
 
         <lang.findUsagesProvider language="yaml" implementationClass="org.zalando.intellij.swagger.reference.usage.YamlFindUsagesProvider"/>
