+++ src/test/resources/testing/rename/json/ref_in_referenced_file_with_caret.json
@@ -1,0 +1,7 @@
+{
+  "Pet1": {
+    "$ref": "#/Pet2<caret>"
+  },
+  "Pet2": {
+  }
+}
