+++ src/test/resources/testing/rename/json/ref_in_referenced_file_with_caret_after.json
@@ -1,0 +1,7 @@
+{
+  "Pet1": {
+    "$ref": "#/newName<caret>"
+  },
+  "newName": {
+  }
+}
