+++ src/test/resources/testing/rename/json/rename_ref_in_referenced_file.json
@@ -1,0 +1,8 @@
+{
+  "swagger": "2.0",
+  "definitions": {
+    "Definition": {
+      "$ref": "definitions.json#/Pet1"
+    }
+  }
+}
