+++ src/test/resources/testing/rename/yaml/ref_in_referenced_file_with_caret_after.yaml
@@ -1,0 +1,3 @@
+Pet1:
+  $ref: '#/newName<caret>'
+newName:
