+++ src/test/resources/testing/rename/yaml/rename_ref_in_referenced_file.yaml
@@ -1,0 +1,4 @@
+swagger: '2.0'
+definitions:
+  Definition:
+    $ref: 'definitions.yaml#/Pet1'
